Similarly, gathered optimiser tests to test_optimisation
