+++ modules/optimisation/__init__.py
@@ -3,6 +3,8 @@
 from .imbalance import make_weighted_sampler
 from .loss import get_loss_function, FocalLoss
 from .scheduler_factory import normalise_scheduler_config, create_scheduler
+from .optimiser_factory import normalise_optimiser_config, create_optimiser
+
 
 __all__ = [
     "normalise_scheduler_config",
@@ -10,4 +12,6 @@ __all__ = [
     "get_loss_function",
     "FocalLoss",
     "make_weighted_sampler",
+    "normalise_optimiser_config",
+    "create_optimiser",
 ]
